subnet-calculator/frontend-python-flask/pyproject.toml
[project]
name = "subnet-calculator-flask-frontend"
version = "0.1.0"
description = "Modern web frontend for IPv4 subnet calculator with cloud provider support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "playwright>=1.55.0",
    "pytest-playwright>=0.7.1",
]
